backend/requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
httpx
chromadb
PyMuPDF
python-multipart
